refactor(android): rebuild NetworkModule and DefaultHttpAdapter
